requirements.txt
pyautogen
autogen-agentchat
autogen-ext
autogen-core
openai
python-dotenv
tiktoken
autogenstudio
mcp
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
